+++ src/cdn/azext_cdn/aaz/latest/afd/origin_group/_create.py
@@ -22,9 +22,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2025-09-01-preview",
+        "version": "2026-04-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cdn/profiles/{}/origingroups/{}", "2025-09-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cdn/profiles/{}/origingroups/{}", "2026-04-01-preview"],
         ]
     }
 
@@ -85,7 +85,7 @@ class Create(AAZCommand):
             options=["--probe-protocol"],
             arg_group="HealthProbeSettings",
             help="Protocol to use for health probe.",
-            enum={"Grpc": "Grpc", "Http": "Http", "Https": "Https", "NotSet": "NotSet"},
+            enum={"Http": "Http", "Https": "Https", "NotSet": "NotSet"},
         )
         _args_schema.probe_request_type = AAZStrArg(
             options=["--probe-request-type"],
@@ -126,6 +126,11 @@ class Create(AAZCommand):
         authentication.scope = AAZStrArg(
             options=["scope"],
             help="The scope used when requesting token from Microsoft Entra. For example, for Azure Blob Storage, scope could be \"https://storage.azure.com/.default\".",
+        )
+        authentication.token_destination_header = AAZStrArg(
+            options=["token-destination-header"],
+            help="The HTTP request header where the origin authentication token will be placed when forwarding the request to the origin. If not specified, the service will use the `Authorization` header for backward compatibility.",
+            enum={"Authorization": "Authorization", "X-Azure-Authorization": "X-Azure-Authorization"},
         )
         authentication.type = AAZStrArg(
             options=["type"],
@@ -229,7 +234,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-09-01-preview",
+                    "api-version", "2026-04-01-preview",
                     required=True,
                 ),
             }
@@ -265,6 +270,7 @@ class Create(AAZCommand):
             authentication = _builder.get(".properties.authentication")
             if authentication is not None:
                 authentication.set_prop("scope", AAZStrType, ".scope")
+                authentication.set_prop("tokenDestinationHeader", AAZStrType, ".token_destination_header")
                 authentication.set_prop("type", AAZStrType, ".type")
                 authentication.set_prop("userAssignedIdentity", AAZObjectType, ".user_assigned_identity")
 
@@ -372,6 +378,9 @@ class _CreateHelper:
 
         authentication = _schema_afd_origin_group_read.properties.authentication
         authentication.scope = AAZStrType()
+        authentication.token_destination_header = AAZStrType(
+            serialized_name="tokenDestinationHeader",
+        )
         authentication.type = AAZStrType()
         authentication.user_assigned_identity = AAZObjectType(
             serialized_name="userAssignedIdentity",