+++ src/horizondb/azext_horizondb/cluster_commands.py
@@ -7,9 +7,13 @@ from azure.cli.core.commands import CliCommandType
 from azext_horizondb._client_factory import (
     cf_horizondb_clusters,
     cf_horizondb_parameter_groups,
-    cf_horizondb_firewall_rules)
+    cf_horizondb_firewall_rules,
+    cf_horizondb_private_endpoint_connections,
+    cf_horizondb_private_link_resources)
 from azext_horizondb.utils._transformers import (
     table_transform_output)
+from azext_horizondb.utils.validators import (
+    validate_private_endpoint_connection_id)
 
 
 # pylint: disable=too-many-locals, too-many-statements, line-too-long
@@ -33,6 +37,16 @@ def load_command_table(self, _):
     firewall_rule_custom = CliCommandType(
         operations_tmpl='azext_horizondb.commands.firewall_rule_commands#{}',
         client_factory=cf_horizondb_firewall_rules)
+    private_endpoint_commands = CliCommandType(
+        operations_tmpl='azext_horizondb.commands.private_endpoint_commands#{}')
+    horizondb_private_endpoint_connections_sdk = CliCommandType(
+        operations_tmpl='azext_horizondb.vendored_sdks.operations#HorizonDbPrivateEndpointConnectionsOperations.{}',
+        client_factory=cf_horizondb_private_endpoint_connections
+    )
+    horizondb_private_link_resources_sdk = CliCommandType(
+        operations_tmpl='azext_horizondb.vendored_sdks.operations#HorizonDbPrivateLinkResourcesOperations.{}',
+        client_factory=cf_horizondb_private_link_resources
+    )
 
     with self.command_group('horizondb', horizondb_clusters_sdk,
                             custom_command_type=custom_commands,
@@ -60,3 +74,22 @@ def load_command_table(self, _):
         g.custom_command('delete', 'horizondb_firewall_rule_delete')
         g.custom_show_command('show', 'horizondb_firewall_rule_get')
         g.custom_command('list', 'horizondb_firewall_rule_list')
+
+    with self.command_group('horizondb private-endpoint-connection',
+                            horizondb_private_endpoint_connections_sdk,
+                            custom_command_type=private_endpoint_commands,
+                            client_factory=cf_horizondb_private_endpoint_connections) as g:
+        g.custom_command('list', 'horizondb_private_endpoint_connection_list')
+        g.show_command('show', 'get', validator=validate_private_endpoint_connection_id)
+        g.command('delete', 'begin_delete', validator=validate_private_endpoint_connection_id)
+        g.custom_command('approve', 'horizondb_approve_private_endpoint_connection',
+                         validator=validate_private_endpoint_connection_id)
+        g.custom_command('reject', 'horizondb_reject_private_endpoint_connection',
+                         validator=validate_private_endpoint_connection_id)
+
+    with self.command_group('horizondb private-link-resource',
+                            horizondb_private_link_resources_sdk,
+                            custom_command_type=private_endpoint_commands,
+                            client_factory=cf_horizondb_private_link_resources) as g:
+        g.custom_command('list', 'horizondb_private_link_resource_list')
+        g.custom_show_command('show', 'horizondb_private_link_resource_get')